requirements.txt
matplotlib==3.10.7
numpy==2.3.5
pandas==2.3.3
scikit-learn==1.7.2
seaborn==0.13.2
joblib==1.5.2
pyspark==4.0.1
boto3==1.41.2
dotenv==0.9.9
findspark==2.0.1
streamlit-1.51.0